requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
colorama==0.4.4
Pillow==9.1.0
pyserial==3.5
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
six==1.16.0